requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-WTF==1.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.2.0
MarkupSafe==3.0.2
numpy==1.24.3
packaging==24.2
pandas==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.4.1.post1 
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
